nerf/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
torch>=1.9.0
torchvision>=0.10.0
torchaudio>=0.9.0
imageio>=2.9.0
imageio-ffmpeg>=0.4.5
matplotlib>=3.3.0
configargparse>=1.4.0
tensorboard>=2.0.0
tqdm>=4.60.0
opencv-python>=4.5.0
setuptools>=65.0.0
wheel>=0.37.0
numpy>=1.21.0
scipy>=1.7.0
pillow>=8.3.0
PyYAML==6.0.2
# torchsearchsorted - install separately if needed:
# pip install git+https://github.com/aliutkus/torchsearchsortedq